+++ app/src/assets/icons/ArrowRightBlueSquare.js
@@ -1,0 +1,15 @@
+import * as React from "react";
+
+const ArrowRightBlueSquare = () => (
+  <svg width="40" height="40" viewBox="0 0 40 40" fill="none" xmlns="http://www.w3.org/2000/svg">
+    <rect width="40" height="40" transform="translate(0 -0.000976562)" fill="#000091" />
+    <path
+      fillRule="evenodd"
+      clipRule="evenodd"
+      d="M24.172 18.9987L18.808 13.6347L20.222 12.2207L28 19.9987L20.222 27.7767L18.808 26.3627L24.172 20.9987H12V18.9987H24.172Z"
+      fill="white"
+    />
+  </svg>
+);
+
+export default ArrowRightBlueSquare;